Use mocha --compilers to test import runtime
