+++ test/feature/Modules/resources/default-class.js
@@ -1,3 +1,5 @@
+import {assert} from '../../../asserts.js';
+
 export default class C {
   m() {
     return 'm';
