+++ test/feature/Modules/resources/default-function.js
@@ -1,3 +1,5 @@
+import {assert} from '../../../asserts.js';
+
 export default function f() {
   return 123;
 }
