+++ test/feature/Syntax/NoNewLineHereEndOfFile.script.js
