+++ test/feature/TemplateLiterals/resources/m.js
@@ -1,3 +1,5 @@
+
+import {assert} from '../../../asserts.js';
 import {f} from './f.js';
 
 assert.equal('a', (f `a`)[0][0]);
