+++ test/feature/TemplateLiterals/resources/n.js
@@ -1,3 +1,4 @@
+import {assert} from '../../../asserts.js';
 import {f} from './f.js';
 
 assert.equal('b', (f `b`)[0][0]);
