+++ test/featureTestRunner.js
@@ -18,63 +18,10 @@ import {NodeTraceurTestRunner} from './modular/NodeTraceurTestRunner.js';
 import {BrowserTraceurTestRunner} from './modular/BrowserTraceurTestRunner.js';
 import {Options} from '../src/Options.js';
 import {ModuleStore} from '../src/loader/ModuleStore.js';
+import parseProlog from '../src/util/parseProlog.js';
 
 import {assert} from './asserts.js';
 export * from './asserts.js';
-
-function forEachPrologLine(s, f) {
-  let inProlog = true;
-  for (let i = 0; inProlog && i < s.length; ) {
-    let j = s.indexOf('\n', i);
-    if (j == -1)
-      break;
-    if (s[i] === '/' && s[i + 1] === '/') {
-      let line = s.slice(i, j);
-      f(line);
-      i = j + 1;
-    } else {
-      inProlog = false;
-    }
-  }
-}
-
-export function parseProlog(source) {
-  let returnValue = {
-    onlyInBrowser: false,
-    skip: false,
-    get shouldHaveErrors() {
-      return this.expectedErrors.length !== 0;
-    },
-    expectedErrors: [],
-    async: false
-  };
-  forEachPrologLine(source, (line) => {
-    let m;
-    if (line.indexOf('// Only in browser.') === 0) {
-      returnValue.onlyInBrowser = true;
-    } else if (line.indexOf('// Skip') === 0) {
-      if (line.indexOf('// Skip.') === 0) {
-        returnValue.skip = true;
-      } else {
-        // eval remainder of line.
-        let skip = false;
-        try {
-            skip = eval(line.slice('// Skip'.length));
-        } catch (ex) {
-            skip = true;
-        }
-        returnValue.skip = !!skip;
-      }
-    } else if (line.indexOf('// Async.') === 0) {
-      returnValue.async = true;
-    } else if ((m = /\/\ Options:\s*(.+)/.exec(line))) {
-      returnValue.traceurOptions = traceur.util.CommandOptions.fromString(m[1]);
-    } else if ((m = /\/\/ Error:\s*(.+)/.exec(line))) {
-      returnValue.expectedErrors.push(m[1]);
-    }
-  });
-  return returnValue;
-}
 
 assert.type = function (actual, type) {
   assert.typeOf(actual, type.name);
